Not trigger ActiveRecord loading on requiring gem
When this gem required ActiveRecord::Base triggered to load and make connections to database. This should not be due to (for example) failing assets:precomplie in Heroku (https://devcenter.heroku.com/articles/rails-asset-pipeline#troubleshooting), where build phase and run phase are separated and database is not available on build phase.